src/main/java/Calculator.java
public class Calculator {

    public static void main(String[]args){
        System.out.println("Hello world");
    }
    public double sum(double num1, double num2) {
        return num1 + num2;
    }

    public double subtraction(double num1, double num2) {
        return num1 - num2;
    }

    public double multiplication(double num1, double num2) {
        return num1 * num2;
    }

    public double division(double num1, double num2) {
        if (num2 != 0) {
            return num1 / num2;
        } else {
            throw new ArithmeticException("ArithmeticException");
        }
    }
}
